requirements.txt
qdrant-client
transformers 
torch 
numpy
pillow